adding authentication

+++ part4/controllers/blogs.js
@@ -1,7 +1,16 @@
+const jwt = require('jsonwebtoken');
 const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
 const User = require('../models/user');
+const config = require('../utils/config')
 
+const getTokenFrom = request => {
+  const authorization = request.get('authorization')
+  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+    return authorization.substring(7)
+  }
+  return null
+}
 
 blogsRouter.get('/', async (request, response, next) => {
  try{
@@ -16,8 +25,12 @@ blogsRouter.get('/', async (request, response, next) => {
 
 blogsRouter.post('/', async (request, response, next) => {
   const body = request.body
-
-  const user = await User.findById(body.userId)
+  const token = getTokenFrom(request)
+  const decodedToken = jwt.verify(token, config.SECRET)
+  if(!decodedToken.id){
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
   const blog = new Blog({
     title: body.title,
     author: body.author,
@@ -38,12 +51,23 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   const {id} = request.params
+  const token = getTokenFrom(request)
+  const decodedToken = jwt.verify(token, config.SECRET)
+  if(!decodedToken.id){
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
   try{
-  const deletedBlog = await Blog.findByIdAndDelete(id)
-  response.status(200).send({
+  //check for wrong user(tomorrow)
+  const blog = await Blog.findById(id)
+  if(blog.user.toString() === decodedToken.id.toString()){
+    const deletedBlog = await Blog.findByIdAndDelete(id);
+    response.status(200).send({
   message: "Resourse deleted successfully",
   deletedId: deletedBlog._id
   })
+  }else{
+    return response.status(401).json({ error: 'action not allowed' })
+  }
   }catch(error){
   next(error)
   }
